Trabalho-1-estacionamento/requirements.txt
sockets
requests
network
flask
RPi.GPIO